backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic[email]
python-dateutil
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosmtplib
python-dotenv
